// apps/ytmp3-clone-4/src/features/downloader/state/types.ts
export type FormatType = 'mp3' | 'mp4';
export type AudioFormatType = 'mp3' | 'wav' | 'm4a' | 'opus' | 'ogg';